delete walletAccount modal and its actions.
