+++ wdio/features/AddressFlow.feature
@@ -6,7 +6,7 @@ Feature: Adding addresses to contacts via address book modal and the contacts sc
   The contacts you saved on network A does not carry over to network B
 
   Scenario: Import account and navigate to Send
-    Given I import wallet using seed phrase "fold media south add since false relax immense pause cloth just raven"
+    Given I have imported my wallet
     And I tap No Thanks on the Enable security check screen
     And I tap No thanks on the onboarding welcome tutorial
     And On the Main Wallet view I tap "Send"
