+++ wdio/features/step-definitions/common-steps.js
@@ -48,6 +48,20 @@ Given(/^I have created my wallet$/, async () => { // should be in a common step 
 
 });
 
+Given(/^I import wallet using seed phrase "([^"]*)?"/, async (phrase) => {
+  const setTimeout = 50000;
+  await driver.pause(setTimeout);
+  await WelcomeScreen.clickGetStartedButton();
+  await OnboardingScreen.clickImportWalletButton();
+  await MetaMetricsScreen.swipeUp();
+  await MetaMetricsScreen.tapIAgreeButton();
+  const validAccount = Accounts.getValidAccount();
+  await ImportFromSeedScreen.typeSecretRecoveryPhrase(phrase);
+  await ImportFromSeedScreen.typeNewPassword(validAccount.password);
+  await ImportFromSeedScreen.typeConfirmPassword(validAccount.password);
+  await ImportFromSeedScreen.clickImportButton();
+});
+
 Given(/^I tap No thanks on the onboarding welcome tutorial/, async () => {
   await OnboardingWizardModal.isVisible();
   const setTimeout = 1500;
