+++ wdio/features/step-definitions/import-wallet-via-private-key.steps.js
@@ -2,7 +2,7 @@ import {When, Then } from '@wdio/cucumber-framework';
 import AccountListComponent from '../screen-objects/AccountListComponent';
 import ImportAccountScreen from '../screen-objects/ImportAccountScreen';
 import ImportSuccessScreen from '../screen-objects/ImportSuccessScreen';
-import WalletAccountModal from "../screen-objects/Modals/WalletAccountModal.js";
+import WalletMainScreen from '../screen-objects/WalletMainScreen';
 
 When(/^I tap on Import an account/, async () => {
     await driver.pause(setTimeout);  
@@ -31,7 +31,7 @@ Then(/^The account is imported/, async () => {
 
 Then(/^I am on the imported account/, async () => {
     await driver.pause(2500);  
-    WalletAccountModal.isAccountNameLabelEqualTo("Account 3") // this can be better
+    WalletMainScreen.isAccountNameLabelEqualTo("Account 3") // this can be better
 });
 
 Then(/^I should see an error (.*)/, async (errorMessage) => {
